27756 - Fix Slot Elicitation Bug
🐘

This PR fixes a bug that occurs when you elicit a slot. The slot is
elicited once more then the set value of maximum_elicitations.

e.g by default this value is set as zero, however the slot will still
elicit once. If you set this value as 1 the slot will elicit twice

- Set the version to 7.0.0 (breaking change)
- Updated the changelog
- Fixed the bug and updated tests

https://dev.azure.com/AMA-Ent/AMA-Ent/_boards/board/t/Red%20Team/Stories/?workitem=27756